create tests
Created unit tests for basic inputs
